refactor(mespapiers): Move Empty inside HomeLayout/Content
and simplify Content api
